+++ database/migrations/2020_04_04_000004_create_records_table.php
@@ -4,7 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateTasksTable extends Migration
+class CreateRecordsTable extends Migration
 {
     /**
      * Run the migrations.
@@ -13,27 +13,20 @@ class CreateTasksTable extends Migration
      */
     public function up()
     {
-        Schema::create('tasks', function (Blueprint $table) {
+        Schema::create('records', function (Blueprint $table) {
             $table->bigIncrements('id');
 
             $table->foreignId('user_id')->nullable()
             ->references(user_model()->getKeyName())
             ->on(user_model()->getTable());
 
-            $table->foreignId('priority_id')->nullable()
-            ->references('id')
-            ->on('priorities');
+            $table->integer('recordable_id');
+            $table->string('recordable_type');
 
-            $table->string('title');
-            $table->text('description')->nullable();
-            $table->text('notes')->nullable();
-            $table->boolean('complete')->nullable()->default(false);
+            $table->timestamp('time_from');
+            $table->timestamp('time_to');
 
-            // TODO:
-            // cost
-            // budget
-
-            // time allocated
+            $table->text('comments')->nullable();
             $table->timestamps();
         });
     }
@@ -45,6 +38,6 @@ class CreateTasksTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('tasks');
+        Schema::dropIfExists('records');
     }
 }