+++ src/Models/Task.php
@@ -3,8 +3,6 @@
 namespace Chriscreates\Projects\Models;
 
 use Chriscreates\Projects\Traits\HasPriority;
-use Chriscreates\Projects\Traits\HasStatus;
-use Chriscreates\Projects\Traits\IsMeasurable;
 use Chriscreates\Projects\Traits\IsRecordable;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -13,7 +11,7 @@ use Illuminate\Database\Eloquent\Relations\MorphToMany;
 
 class Task extends Model
 {
-    use IsMeasurable, IsRecordable, HasStatus, HasPriority;
+    use IsRecordable, HasPriority;
 
     protected $table = 'tasks';
 
@@ -23,20 +21,13 @@ class Task extends Model
 
     public $timestamps = true;
 
-    protected $dates = [
-        'started_at',
-        'delivered_at',
-        'expected_at',
+    protected $casts = [
+        'complete' => 'bool',
     ];
 
     public function creator() : BelongsTo
     {
         return $this->belongsTo(get_class(user_model()));
-    }
-
-    public function status() : BelongsTo
-    {
-        return $this->belongsTo(Status::class);
     }
 
     public function priority() : BelongsTo
@@ -52,5 +43,22 @@ class Task extends Model
     public function records() : MorphMany
     {
         return $this->morphMany(Record::class, 'recordable');
+    }
+
+    public function assignToProject(Project $project) : void
+    {
+        $this->projects()->save($project);
+
+        $this->refresh();
+    }
+
+    public function markCompletion(bool $bool) : void
+    {
+        $this->update(['complete' => $bool]);
+    }
+
+    public function completed() : bool
+    {
+        return $this->attributes['complete'];
     }
 }